feat(shared-features-social): restructure feature and add Nx boilerplate
- Restructure social feature to follow feature.definition.md architecture
- Create src/ directory with application, infrastructure, and presentation layers
- Add missing social domain exports in libs/domains/catalog/social
- Fix all import paths to use proper domain library imports
- Create missing social provider port interface
- Move files to proper architectural layers
- Add @domains/catalog/social and @portals/shared/features/social aliases to root tsconfig
- Add complete Nx boilerplate: tsconfig files, project.json, jest.config.ts
- Follow established domain-driven architecture pattern with proper layer separation
