+++ apps/portals/shared/features/social/project.json
@@ -1,0 +1,38 @@
+{
+  "name": "portals-shared-features-social",
+  "$schema": "../../../../../node_modules/nx/schemas/project-schema.json",
+  "projectType": "library",
+  "prefix": "social",
+  "sourceRoot": "apps/portals/shared/features/social",
+  "tags": ["type:lib", "scope:shared", "scope:features"],
+  "targets": {
+    "build": {
+      "executor": "@nx/js:tsc",
+      "outputs": ["{options.outputPath}"],
+      "options": {
+        "outputPath": "dist/apps/portals/shared/features/social",
+        "main": "apps/portals/shared/features/social/src/index.ts",
+        "tsConfig": "apps/portals/shared/features/social/tsconfig.lib.json",
+        "assets": ["apps/portals/shared/features/social/*.md"]
+      }
+    },
+    "test": {
+      "executor": "@nx/jest:jest",
+      "outputs": ["{workspaceRoot}/coverage/{projectRoot}"],
+      "options": {
+        "jestConfig": "apps/portals/shared/features/social/jest.config.ts",
+        "passWithNoTests": true
+      },
+      "configurations": {
+        "ci": {
+          "ci": true,
+          "codeCoverage": true
+        }
+      }
+    },
+    "lint": {
+      "executor": "@nx/eslint:lint",
+      "outputs": ["{options.outputFile}"]
+    }
+  }
+}
