+++ apps/portals/shared/features/social/src/application/index.ts
@@ -1,0 +1,2 @@
+export * from './ports/social-provider.port';
+export * from './social.service';
