+++ apps/portals/shared/features/social/src/application/ports/social-provider.port.ts
@@ -1,0 +1,10 @@
+import { InjectionToken } from "@angular/core";
+import { Observable } from "rxjs";
+import { Result } from "@standard";
+import { SocialDto } from "@domains/catalog/social";
+
+export const SOCIALS_PROVIDER = new InjectionToken<ISocialsProvider>('SOCIALS_PROVIDER_PORT');
+
+export interface ISocialsProvider {
+  getSocials(): Observable<Result<SocialDto[], Error>>;
+}
