+++ apps/portals/shared/features/social/src/index.ts
@@ -1,0 +1,3 @@
+export * from './application/social.service';
+export * from './social.providers';
+export * from './presentation/social-list-container.directive';
