+++ apps/portals/shared/features/social/src/infrastructure/index.ts
@@ -1,0 +1,1 @@
+export * from './social-api.service';
