+++ apps/portals/shared/features/social/src/presentation/index.ts
@@ -1,0 +1,1 @@
+export * from './social-list-container.directive';
