+++ libs/domains/catalog/social/application/index.ts
@@ -1,0 +1,2 @@
+export * from './models/social.dto';
+export * from './ports/social-provider.port';
