+++ tsconfig.base.json
@@ -68,6 +68,7 @@
       "@domains/catalog/entry": ["libs/domains/catalog/entry/application/index.ts"],
       "@domains/catalog/platform": ["libs/domains/catalog/platform/index.ts"],
       "@domains/catalog/compatibility": ["libs/domains/catalog/compatibility/application/index.ts"],
+      "@domains/catalog/social": ["libs/domains/catalog/social/application/index.ts"],
       "@domains/customer/my-profile": ["libs/domains/customer/my-profile/application/index.ts"],
       "@portals/shared/features/categories": ["apps/portals/shared/features/categories/src/index.ts"],
       "@portals/shared/features/listing": ["apps/portals/shared/features/listing/src/index.ts"],
